extract matcher and generator entirely
