+++ tests/GeneratorTest.php
@@ -5,16 +5,14 @@ use ArrayObject;
 
 class GeneratorTest extends \PHPUnit_Framework_TestCase
 {
-    protected $factory;
-    protected $server;
+    protected $routes;
     protected $generator;
 
     protected function setUp()
     {
         parent::setUp();
-        $this->generator = new Generator();
-        $this->factory = new RouteFactory();
-        $this->server = $_SERVER;
+        $this->routes = new RouteCollection(new RouteFactory());
+        $this->generator = new Generator($this->routes);
     }
 
     /**
@@ -22,7 +20,7 @@ class GeneratorTest extends \PHPUnit_Framework_TestCase
      */
     public function testGenerateControllerAsClosureIssue19()
     {
-        $route = $this->factory->newInstance('/blog/{id}/edit')
+        $this->routes->add('issue19', '/blog/{id}/edit')
             ->setTokens(array(
                 'id' => '([0-9]+)',
             ))
@@ -35,24 +33,30 @@ class GeneratorTest extends \PHPUnit_Framework_TestCase
                 'format' => '.html',
             ));
 
-        $url = $this->generator->generate($route, array('id' => 42, 'foo' => 'bar'));
+        $url = $this->generator->generate('issue19', array('id' => 42, 'foo' => 'bar'));
         $this->assertEquals('/blog/42/edit', $url);
     }
 
     public function testGenerate()
     {
-        $route = $this->factory->newInstance('/blog/{id}/edit')
+        $this->routes->add('test', '/blog/{id}/edit')
             ->setTokens(array(
                 'id' => '([0-9]+)',
             ));
 
-        $url = $this->generator->generate($route, array('id' => 42, 'foo' => 'bar'));
+        $url = $this->generator->generate('test', array('id' => 42, 'foo' => 'bar'));
         $this->assertEquals('/blog/42/edit', $url);
+    }
+
+    public function testGenerateMissing()
+    {
+        $this->setExpectedException('Aura\Router\Exception\RouteNotFound');
+        $this->generator->generate('no-such-route');
     }
 
     public function testGenerateWithClosure()
     {
-        $route = $this->factory->newInstance('/blog/{id}/edit')
+        $this->routes->add('test', '/blog/{id}/edit')
             ->setTokens(array(
                 'id' => '([0-9]+)',
             ))
@@ -60,31 +64,31 @@ class GeneratorTest extends \PHPUnit_Framework_TestCase
                 $data['id'] = 99;
             });
 
-        $url = $this->generator->generate($route, array('id' => 42, 'foo' => 'bar'));
+        $url = $this->generator->generate('test', array('id' => 42, 'foo' => 'bar'));
         $this->assertEquals('/blog/99/edit', $url);
     }
 
     public function testGenerateWithCallback()
     {
-        $route = $this->factory->newInstance('/blog/{id}/edit')
+        $this->routes->add('test', '/blog/{id}/edit')
             ->setTokens(array(
                 'id' => '([0-9]+)',
             ))
             ->setGenerateCallable(array($this, 'callbackForGenerate'));
 
-        $url = $this->generator->generate($route, array('id' => 42, 'foo' => 'bar'));
+        $url = $this->generator->generate('test', array('id' => 42, 'foo' => 'bar'));
         $this->assertEquals('/blog/99/edit', $url);
     }
 
     public function testGenerateWithWildcard()
     {
-        $route = $this->factory->newInstance('/blog/{id}')
+        $this->routes->add('test', '/blog/{id}')
             ->setTokens(array(
                 'id' => '([0-9]+)',
             ))
             ->setWildcard('other');
 
-        $url = $this->generator->generate($route, array(
+        $url = $this->generator->generate('test', array(
             'id' => 42,
             'foo' => 'bar',
             'other' => array(
@@ -98,9 +102,9 @@ class GeneratorTest extends \PHPUnit_Framework_TestCase
 
     public function testGenerateWithOptional()
     {
-        $route = $this->factory->newInstance('/archive/{category}{/year,month,day}');
+        $this->routes->add('test', '/archive/{category}{/year,month,day}');
 
-        $url = $this->generator->generate($route, array(
+        $url = $this->generator->generate('test', array(
             'category' => 'foo',
             'year' => '1979',
             'month' => '11',
@@ -116,39 +120,39 @@ class GeneratorTest extends \PHPUnit_Framework_TestCase
 
     public function testGenerateOnFullUri()
     {
-        $route = $this->factory->newInstance('http://google.com/?q={q}', 'google-search')
+        $this->routes->add('test', 'http://google.com/?q={q}', 'google-search')
             ->setRoutable(false);
 
-        $actual = $this->generator->generate($route, array('q' => "what's up doc?"));
+        $actual = $this->generator->generate('test', array('q' => "what's up doc?"));
         $expect = "http://google.com/?q=what%27s%20up%20doc%3F";
         $this->assertSame($expect, $actual);
     }
 
     public function testGenerateRFC3986()
     {
-        $route = $this->factory->newInstance('/path/{string}', 'rfc3986')
+        $this->routes->add('test', '/path/{string}', 'rfc3986')
             ->setRoutable(false);
 
         // examples taken from http://php.net/manual/en/function.rawurlencode.php
-        $actual = $this->generator->generate($route, array('string' => 'foo @+%/'));
+        $actual = $this->generator->generate('test', array('string' => 'foo @+%/'));
         $expect = '/path/foo%20%40%2B%25%2F';
         $this->assertSame($actual, $expect);
 
-        $actual = $this->generator->generate($route, array('string' => 'sales and marketing/Miami'));
+        $actual = $this->generator->generate('test', array('string' => 'sales and marketing/Miami'));
         $expect = '/path/sales%20and%20marketing%2FMiami';
         $this->assertSame($actual, $expect);
     }
 
     public function testGenerateRaw()
     {
-        $route = $this->factory->newInstance('/{vendor}/{package}/{file}');
+        $this->routes->add('test', '/{vendor}/{package}/{file}');
         $data = array(
             'vendor' => 'vendor+name',
             'package' => 'package+name',
             'file' => 'foo/bar/baz.jpg',
         );
         $raw = array('file');
-        $actual = $this->generator->generateRaw($route, $data);
+        $actual = $this->generator->generateRaw('test', $data);
         $expect = '/vendor+name/package+name/foo/bar/baz.jpg';
         $this->assertSame($actual, $expect);
     }
